Fix crash in tracing code.
Bug: http://b/29358536

When we switched to standalone bcc, we accidentally removed the setters
for this name field. It is simplest to just use the slot number for
fixing this today, and revisit how to populate a proper name later on,
since it would technically require vendor driver updates.

(cherry picked from commit d969fdc55bc82d48242be3c6193a6548137e9579)

Change-Id: I38a67a10d960d8fd2dddd03dc5155e3781f390d8